zsh_remove_unnecessary_plugins works as expected
Everything works on the first run because when plugins are downloaded they are added to the PLUGINS array. However, on subsequent runs, that PLUGINS variable is reset and only gets updated for newly installed plugins. When the remove_unnecessary_plugins function runs, it compares the plugins you have with the empty PLUGINS variable, sees you have nothing set, and tries to remove all plugins. I added code that updates the PLUGINS variable whether you already have the plugin or not (the former didn't do this)