Made neovim standard program use lts function
This commit is mainly to show how the lts function will be used to output a package and its latest version from a specific repo. This meant to be only for Fedora (as it uses dnf). I added this functionality because Fedora has two main versions of neovim (from the fedora and updates repos), and I only want the newer version. Some distros like arch only have one, so you won't need to use these functions