Changed utils.sh file
Needed to add some tweaking so that the functions work as intended with their given parameters. Moved linux_distros variable into utils.sh because it doesn't make sense for it to be exposed to the user in the software_install.sh file. Moved installation_method variable into install_programs function in utils.sh because it should only evaluate when the function is called not every time.