dotfile_setup only runs if there are directories
Without this, if the dotfiles directory had no sub-directories, it would throw an error because it's trying to find the setup.sh file in */ which is not a thing. While not very practical since once you run the install script, while directories will be moved, you won't have a need to run it again. This just helps with debugging after the first script run.